Update old.ts
I sliced the class Car and did the classes MusicPlayer and Engine.

In the class MusicPlayer I put the methods:
-> get musicLevel()
-> set musicLevel()
-> turnMusicOn()
-> turnMusicOff()

I also put the variables:
-> private _musicLevel
-> private _oldMusicLevel
-------------------------------------------
In the class Engine I put the methods:
-> get miles()
-> get fuel()
-> addFuel()
-> get engineStatus()
-> turnEngineOn()
-> turnEngineOff()
-> drive()

I also put the constructor and the variables:
-> _engineStatus
-> _fuel
-> _miles
-> MAXIMUM_FUEL_CAPACITY
-> FUEL_MILEAGE

In the class Car I just instantiated the classes MusicPlayer and Engine.